Much needed documentation update + rewrite of the "help" RPC method
Rewrite of the "help" method to be more human-friendly. A special case
for the "help" method has been introduced in the socket protocol: it
doesn't require to enter the message terminaison string (#end#): if
you're using telnet to discuss with the knowledge base and you're lost,
jsut type "help" + return to get started! Lovely!