Common Ancestor algorithm implemented in DiffModule.
First actual work on the diff and similarites module. This first step
implements a algorithm to find common ancestor types for two given
concept.

The mathematical representation of what we are looking for is:
Let's consider a,b instances of classes A and B:
We want S2 such as:
For S1 = superclasses(A) ∩ superclasses(B),
S2 = { c / c ∈ S1 AND subclasses(c) ∩ S1 = ∅ }

(Sorry for people not using UTF-8, but we live in 2009, don't we? ;-) )