Major rewrite of the events system (first part)
The new event system rely on an updated version of "watchers" (ie patterns
that trigger events when they are matched in the ontology).

This watchers are exposed by classes that want to register event. These classes
must implement the IEventsProvider interface.

Finally, once triggered, events are consummed by classes that implement the
IEventConsumer interface.

More detail in the documentation (package.html) or in the unittests (EventsTest.java)