External event registration and triggering is back!
The example below registers an trigger upon new instances of humans.
It can be tested with telnet:

 > registerEvent
 > new_instance //event type
 > on_true      //triggering type
 > [Human]
 > #end#

The server answer ok, with a unique ID bound to this so-called "event watcher":

 > ok
 > 4565-4587995-112355-21446
 > #end#

If a new human (for instance, "ramses") appears in the ontology, the server
send this kind of notification:

 > event
 > 4565-4587995-112355-21446
 > [ramses]
 > #end#

The API is mostly working, but non-blocking server sockets are required
to get the expected behaviours.