Improved the lookup RPC method to return *all* concepts that have a common label.
This should solve bug 114.
This modification revealed a new, strange bug (Pellet bug?):
If these 2 stmts are added:
KingKong rdfs:subClassOf Monkey
KingKong rdfs:label "king kong"

Jena would throw a conversion exception (*in strict mode*).
If not in strict mode, KingKong is inferred to be both a class and and instance.

Cf comments in testLookup() in test suite to reproduce the bug.

Some other minor enhancements in this commit, like checking that statements are not
null before adding them.