Added update() RPC method
This method takes a statement with a *functional* predicate (ie, a predicate that
can have only one value), and replaces a previous statement with the same subject
and predicate, ie, it updates the value of this property.

If the predicate is unknown or not a functional predicate, update() behaves like
add().