Hopefully fixed the ugly concurrency issues in oro-server
The problem was: since Pellet, the reasoner, is not thread-safe, we want to be sure that we control when it is
triggered. The only way is to classify the ontology as soon as we change it. In the previous version, I wanted
to be clever and classify the ontology only before queries (to avoid classification each time a statement is added).
But I was not clever enough: classification is required in a lot of different places (to build to symbol lookup table,
to process events,etc...). So it's easier to classify as soon as the model change.

Of course, it is costly. So better group the add and remove requests!
Background on Pellet and concurrency: http://clarkparsia.com/pellet/faq/jena-concurrency/