The robot ID is now set in the config file
In the previous situation, one would set the robot ID in the scenario ontology, with
the relation 'sameAs myself'. But this caused trouble when new agents appeared since
they would be 'sameAs myself' and 'sameAs robotid' in the new model.

The best way to avoid it seems to define the robot id outside of the ontologies.