Removed all locking of the RDF model
Since all operations are now taking place in the main thread, no locking
is required anymore. Yeah.

The only operation that should not be possible for now is 'MemoryManagement'.
Don't use it for now (set the option to false in your config file --
default value).

Known bug: Uncatched exceptions are expected when the ontology become
inconsistent.

While here, checkConsistency now returns a boolean instead of throwing
an exception when the ontology is inconsistent.